Crate tokio_buf

Asynchronous stream of bytes.

This crate contains the BufStream trait and a number of combinators for this trait. The trait is similar to Stream in the futures library, but instead of yielding arbitrary values, it only yields types that implement Buf (i.e, byte collections).

Modules

util
Types and utilities for working with BufStream.

Structs

SizeHint
A BufStream size hint

Traits

BufStream
An asynchronous stream of bytes.

BufStreamExt
An extension trait for BufStream's that provides a variety of convenient adapters.