Module tokio_buf::util

Types and utilities for working with BufStream.

Modules

error
Error types

Structs

Chain
A buf stream that sequences two buf streams together.

Collect
Consumes a buf stream, collecting the data into a single byte container.

IntoStream
Converts a BufStream into a Stream.

Limit
Limits the stream to a maximum amount of data.

Traits

BufStreamExt
An extension trait for BufStream's that provides a variety of convenient adapters.

FromBufStream
Conversion from a BufStream.

Functions

iter
Converts an Iterator into a BufStream which is always ready to yield the next value.

stream
Converts a Stream of Buf types into a BufStream.