Function tokio_buf::util::stream

pub fn stream<T>(stream: T) -> FromStream<T> where
    T: Stream,
    T::Item: Buf, 

Converts a Stream of Buf types into a BufStream.

While Stream and BufStream are very similar, they are not identical. The stream function returns a BufStream that is backed by the provided Stream type.